webhooks: remove go1.12 support
